+++ cisco_aci/datadog_checks/cisco_aci/exceptions.py
@@ -13,3 +13,7 @@ class APIConnectionException(APIException):
 
 class APIParsingException(APIException):
     pass
+
+
+class ConfigurationException(Exception):
+    pass