+++ etcd/tests/common.py
@@ -1,0 +1,11 @@
+# (C) Datadog, Inc. 2018
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+import os
+
+from datadog_checks.dev import get_docker_hostname
+
+HERE = os.path.dirname(os.path.abspath(__file__))
+HOST = get_docker_hostname()
+PORT = '2379'
+MAIN_URL = 'http://{}:{}'.format(HOST, PORT)